Rename: 파일명 및 경로 수정, 불필요한 코드 제거'
::wq
